feat(charts): wire all-years slider bound into SimpleView and DetailedView

+++ app/src/components/Charts/DetailedCharts/StreamPerHour/StreamPerHour.sql
@@ -1,7 +1,7 @@
 select
     hour::int as hour,
-    coalesce(count_streams, 0)::int as count_streams,
-    coalesce(ms_played, 0)::int as ms_played
+    coalesce(count_streams, 0)::double as count_streams,
+    coalesce(ms_played, 0)::double as ms_played
 from (select unnest(range(24)) as hour)
 left join (
     select
